+++ src/main/java/org/ojalgo/optimisation/convex/ConvexSolver.java
@@ -45,7 +45,9 @@ import org.ojalgo.matrix.store.MatrixStore;
 import org.ojalgo.matrix.store.PhysicalStore;
 import org.ojalgo.matrix.store.R064Store;
 import org.ojalgo.matrix.task.iterative.ConjugateGradientSolver;
+import org.ojalgo.matrix.task.iterative.JacobiPreconditioner;
 import org.ojalgo.matrix.task.iterative.IterativeSolverTask;
+import org.ojalgo.matrix.task.iterative.Preconditioner;
 import org.ojalgo.optimisation.Expression;
 import org.ojalgo.optimisation.ExpressionsBasedModel;
 import org.ojalgo.optimisation.GenericSolver;
@@ -368,6 +370,7 @@ public abstract class ConvexSolver extends GenericSolver {
         private boolean myCombinedScaleFactor = true;
         private boolean myExtendedPrecision = false;
         private NumberContext myIterativeAccuracy = NumberContext.of(10, 14).withMode(RoundingMode.HALF_DOWN);
+        private Supplier<Preconditioner> myIterativePreconditioner = JacobiPreconditioner::new;
         private Supplier<IterativeSolverTask> myIterativeSolver = ConjugateGradientSolver::new;
         private double mySmallDiagonal = RELATIVELY_SMALL + MACHINE_EPSILON;
         private Function<Structure2D, MatrixDecomposition.Solver<Double>> mySolverGeneral = LU.R064::make;
@@ -447,10 +450,24 @@ public abstract class ConvexSolver extends GenericSolver {
             return this;
         }
 
-        /** Returns the configured iterative solver */
+        public Configuration iterative(final Supplier<IterativeSolverTask> solver, final Supplier<Preconditioner> preconditioner,
+                final NumberContext accuracy) {
+            Objects.requireNonNull(solver);
+            Objects.requireNonNull(preconditioner);
+            Objects.requireNonNull(accuracy);
+            myIterativeSolver = solver;
+            myIterativePreconditioner = preconditioner;
+            myIterativeAccuracy = accuracy;
+            return this;
+        }
+
+        /**
+         * Returns a new iterative solver instance configured with the current accuracy, maximum iterations,
+         * and preconditioner settings.
+         */
         public IterativeSolverTask newIterativeSolver(final int maxIterations) {
             IterativeSolverTask retVal = myIterativeSolver.get();
-            retVal.configurator().accuracy(myIterativeAccuracy).iterations(maxIterations);
+            retVal.configurator().accuracy(myIterativeAccuracy).iterations(maxIterations).preconditioner(myIterativePreconditioner.get());
             return retVal;
         }